requirements.txt
absl-py==0.9.0
astor==0.8.0
bleach==3.1.0
captcha==0.3
certifi==2020.4.5.1
chardet==3.0.4
Click==7.1.2
cycler==0.10.0
easydict==1.8
Flask==1.1.1
gast==0.2.2
grpcio==1.32.0
html5lib==0.9999999
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.17.0
olefile==0.46
Pillow==6.2.1
protobuf==3.13.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
tensorboard==2.2.1
tensorflow==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.16.1
fsspec==0.6.2
gcsfs==0.4.0
google-cloud-error-reporting==0.30.0
google-cloud-storage==1.20.0
